Fixed warnings

+++ src/scenemanager.cpp
@@ -24,7 +24,7 @@ void SceneManager::operator <<(const YAML::Node &scenes)
 
 const std::shared_ptr<Scene> SceneManager::current()
 {
-  if (current_ >= 0 || current_ < size()) {
+  if (current_ >= 0 || current_ < static_cast<int>(size())) {
     return at(current_);
   } else {
     return nullptr;
@@ -36,7 +36,7 @@ const std::shared_ptr<Scene> SceneManager::next()
   ++current_;
   assert(current_ >= 0);
 
-  if (current_ < size()) {
+  if (current_ < static_cast<int>(size())) {
     return at(current_);
   } else {
     return nullptr;
